Package reorganizations. Small refactoring.
