+++ src/main/java/com/github/prominence/openweathermap/api/request/ResponseMapper.java
@@ -20,35 +20,13 @@
  * SOFTWARE.
  */
 
-package com.github.prominence.openweathermap.api.impl;
+package com.github.prominence.openweathermap.api.request;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.stream.Collectors;
+import java.util.List;
 
-public class RequestUrlBuilder {
+public interface ResponseMapper<T> {
 
-    private StringBuilder builder = new StringBuilder();
-    private Map<String, Object> requestParameters = new HashMap<>();
+    T getSingle(String json);
 
-    public RequestUrlBuilder(String baseUrl) {
-        builder.append(baseUrl);
-    }
-
-    void append(String value) {
-        builder.append(value);
-    }
-
-    public void addRequestParameter(String key, Object value) {
-        requestParameters.put(key, value);
-    }
-
-    String buildUrl() {
-        final String joinedParameters = requestParameters.entrySet().stream()
-                .map(entry -> entry.getKey() + "=" + entry.getValue())
-                .collect(Collectors.joining("&"));
-        builder.append('?');
-        builder.append(joinedParameters);
-        return builder.toString();
-    }
+    List<T> getList(String json);
 }
