+++ src/main/java/com/github/prominence/openweathermap/api/request/weather/multiple/MultipleResultCurrentWeatherAsyncRequestTerminatorImpl.java
@@ -20,43 +20,47 @@
  * SOFTWARE.
  */
 
-package com.github.prominence.openweathermap.api.impl;
+package com.github.prominence.openweathermap.api.request.weather.multiple;
 
-import com.github.prominence.openweathermap.api.SingleResultCurrentWeatherRequestTerminator;
+import com.github.prominence.openweathermap.api.request.RequestUrlBuilder;
+import com.github.prominence.openweathermap.api.request.weather.CurrentWeatherResponseMapper;
 import com.github.prominence.openweathermap.api.enums.UnitSystem;
 import com.github.prominence.openweathermap.api.model.Weather;
 import com.github.prominence.openweathermap.api.utils.RequestUtils;
 
-public class SingleResultCurrentWeatherRequestTerminatorImpl implements SingleResultCurrentWeatherRequestTerminator {
+import java.util.List;
+import java.util.concurrent.CompletableFuture;
 
-    private RequestUrlBuilder urlBuilder;
-    private UnitSystem unitSystem;
+public class MultipleResultCurrentWeatherAsyncRequestTerminatorImpl implements MultipleResultCurrentWeatherAsyncRequestTerminator {
 
-    SingleResultCurrentWeatherRequestTerminatorImpl(RequestUrlBuilder urlBuilder, UnitSystem unitSystem) {
+    private final RequestUrlBuilder urlBuilder;
+    private final UnitSystem unitSystem;
+
+    MultipleResultCurrentWeatherAsyncRequestTerminatorImpl(RequestUrlBuilder urlBuilder, UnitSystem unitSystem) {
         this.urlBuilder = urlBuilder;
         this.unitSystem = unitSystem;
     }
 
     @Override
-    public Weather asJava() {
-        return new CurrentWeatherResponseMapper(unitSystem).getSingle(asJSON());
+    public CompletableFuture<List<Weather>> asJava() {
+        return CompletableFuture.supplyAsync(() -> new CurrentWeatherResponseMapper(unitSystem).getList(getRawResponse()));
     }
 
     @Override
-    public String asJSON() {
-        return getRawResponse();
+    public CompletableFuture<String> asJSON() {
+        return CompletableFuture.supplyAsync(this::getRawResponse);
     }
 
     @Override
-    public String asXML() {
+    public CompletableFuture<String> asXML() {
         urlBuilder.addRequestParameter("mode", "xml");
-        return getRawResponse();
+        return CompletableFuture.supplyAsync(this::getRawResponse);
     }
 
     @Override
-    public String asHTML() {
+    public CompletableFuture<String> asHTML() {
         urlBuilder.addRequestParameter("mode", "html");
-        return getRawResponse();
+        return CompletableFuture.supplyAsync(this::getRawResponse);
     }
 
     private String getRawResponse() {
