+++ src/test/java/com/github/prominence/openweathermap/api/ApiTest.java
@@ -22,7 +22,7 @@
 
 package com.github.prominence.openweathermap.api;
 
-import com.github.prominence.openweathermap.api.impl.OpenWeatherMapClient;
+import com.github.prominence.openweathermap.api.request.OpenWeatherMapClient;
 import org.junit.BeforeClass;
 
 public class ApiTest {
