+++ src/test/java/com/github/prominence/openweathermap/api/CurrentWeatherIntegrationTest.java
@@ -27,7 +27,7 @@ import com.github.prominence.openweathermap.api.enums.Language;
 import com.github.prominence.openweathermap.api.enums.UnitSystem;
 import com.github.prominence.openweathermap.api.exception.NoDataFoundException;
 import com.github.prominence.openweathermap.api.exception.InvalidAuthTokenException;
-import com.github.prominence.openweathermap.api.impl.OpenWeatherMapClient;
+import com.github.prominence.openweathermap.api.request.OpenWeatherMapClient;
 import com.github.prominence.openweathermap.api.model.Coordinate;
 import com.github.prominence.openweathermap.api.model.CoordinateRectangle;
 import com.github.prominence.openweathermap.api.model.Weather;
